+++ budget-app/src/pages/Generic/NoPage.jsx
