+++ budget-app/src/pages/Months/Months.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 
-import EditModalButton from './EditModalButton'
+import EditModalButton from '../Generic/EditModalButton'
 import Table from '@mui/material/Table';
 import TableBody from '@mui/material/TableBody';
 import TableCell from '@mui/material/TableCell';
